requirements.txt
pillow==10.4.0
asgiref==3.8.1
Django==5.1
Markdown==3.7
django-markdownx==4.0.7
sqlparse==0.5.1
tinycss2==1.2.1
six==1.16.0
webencodings==0.5.1
bleach==6.1.0
django-markdownify==0.9.5
